requirements.txt
flask
requests
pandas
pyarrow
google-cloud-storage
gunicorn
google-cloud-secret-manager
